requirements.txt
cv2
numpy
openpyxl
pyautogui
pytesseract
easygui
Pillow
opencv-python
